insert triggered campaign campaign message to campaign_messages table
This fixes that triggered campaign e-mails are not reported in ui and so it's also possible to
create another triggered campaign on this first triggered one.

What i don't understand why their is a difference in handling of a triggered campaign and a normal one in sending.

Also i expected that a triggered campaign can only be send once to an user regardless which trigger fires the send.
But currently it's send for every trigger that is defined in a campaign.